bugfixes: reponse last char replaced w/ \0 fixed
sidenote: request log format fixed

+++ src/stdfunc.c
@@ -39,10 +39,10 @@ const char* server_std_gettime()
 
 int server_std_get_sub_reqdata_end(const char* reqdata)
 {
-    return (int) MIN(                          // get min of the following three positions
+    return (int) MIN(                                        // get min of the following three positions
         MIN(
-            32,                                // max 32, position of 32nd char
-            strchr(reqdata, '\n') - reqdata    // position of \n in request
-        ), strchr(reqdata, '\r') - reqdata     // position of \r in request
+           32,                                               // max 32, position of 32nd char
+           (unsigned int) (strchr(reqdata, '\n') - reqdata)  // position of \n in request
+        ), (unsigned int) (strchr(reqdata, '\r') - reqdata)  // position of \r in request
     );
 }
